refactor: group inertial initialization module
