+++ crates/kornia-slam/src/initialization/inertial/factor.rs
