+++ crates/kornia-slam/src/initialization/mod.rs
@@ -1,10 +1,9 @@
 //! Visual and inertial initialization for the SLAM system.
 
-mod imu;
-mod inertial_factor;
+mod inertial;
 mod two_view;
 
-pub use imu::{ImuInitConfig, ImuInitReject, ImuInitResult, ImuInitializer, KeyframeVelocity};
+pub use inertial::{ImuInitConfig, ImuInitReject, ImuInitResult, ImuInitializer, KeyframeVelocity};
 pub use two_view::{
     TwoViewAcceptanceConfig, TwoViewEstimate, TwoViewInitConfig, TwoViewRejectReason,
     try_initialize_two_view,
